refactor(lib): split code
