+++ src/lib/search.ts
@@ -3,12 +3,22 @@
  * MIT Licensed
  */
 
+import { copyGrid, Grid, manhattanDistance } from './grid'
 import { Heap } from './heap'
+import { isValidPosition, MAX_BOUND, Pair, pairSum } from './pair'
 
-export type Pair = [number, number]
-
+// Key
 type Key = string
 
+const keyFromGrid = (grid: Grid) => {
+  let key = ''
+  for (const row of grid) {
+    for (const n of row) key += n.toString()
+  }
+  return key
+}
+
+// Step
 export enum Step {
   Up = '^',
   Left = '<',
@@ -16,9 +26,15 @@ export enum Step {
   Right = '>',
 }
 
-type Row = [number, number, number]
-type Grid = [Row, Row, Row]
+// col X row
+const STEPS: Array<[Step, Pair]> = [
+  [Step.Up, [1, 0]],
+  [Step.Left, [0, 1]],
+  [Step.Down, [-1, 0]],
+  [Step.Right, [0, -1]],
+]
 
+// State
 type State = {
   cost: number
   path: Array<Step>
@@ -28,6 +44,8 @@ type State = {
   grid: Grid
 }
 
+const isFinalState = (state: State) => state.key === '123456780'
+
 export const makeSearchState = (grid: Grid): State => ({
   cost: 0,
   path: [],
@@ -36,25 +54,6 @@ export const makeSearchState = (grid: Grid): State => ({
   zeroPosition: zeroPairFromGrid(grid),
   grid,
 })
-
-const isFinalState = (state: State) => state.key === '123456780'
-
-const MAX_BOUND = 2
-const MIN_BOUND = 0
-const BOUNDS_RANGE: Pair = [MIN_BOUND, MAX_BOUND]
-
-const isWithinRange = ([x, y]: Pair, t: number) => x <= t && t <= y
-
-export const isValidPosition = ([x, y]: Pair) =>
-  isWithinRange(BOUNDS_RANGE, x) && isWithinRange(BOUNDS_RANGE, y)
-
-const keyFromGrid = (grid: Grid) => {
-  let key = ''
-  for (const row of grid) {
-    for (const n of row) key += n.toString()
-  }
-  return key
-}
 
 const zeroPairFromGrid = (grid: Grid): Pair => {
   for (let i = 0; i <= MAX_BOUND; i++) {
@@ -66,28 +65,6 @@ const zeroPairFromGrid = (grid: Grid): Pair => {
   throw Error('Wrong Grid')
 }
 
-const rowFromList = (
-  list: Array<number>,
-  start: number,
-): Row => [list[start], list[start + 1], list[start + 2]]
-export const gridFromList = (
-  list: Array<number>,
-): Grid => [rowFromList(list, 0), rowFromList(list, 3), rowFromList(list, 6)]
-
-// For types
-const copyRow = (row: Row): Row => [row[0], row[1], row[2]]
-const copyGrid = (
-  grid: Grid,
-): Grid => [copyRow(grid[0]), copyRow(grid[1]), copyRow(grid[2])]
-
-// col X row
-const STEPS: Array<[Step, Pair]> = [
-  [Step.Up, [1, 0]],
-  [Step.Left, [0, 1]],
-  [Step.Down, [-1, 0]],
-  [Step.Right, [0, -1]],
-]
-
 type DataStructure = {
   add(state: State): void
   extract(): State
@@ -96,44 +73,19 @@ type DataStructure = {
   cost(state: State): number
 }
 
-export const pairEq = ([a, b]: Pair, [c, d]: Pair) => a == c && b == d
-export const pairSum = (
-  [x1, y1]: Pair,
-  [x2, y2]: Pair,
-): Pair => [x1 + x2, y1 + y2]
-const pairDiff = ([x1, y1]: Pair, [x2, y2]: Pair): Pair => [x1 - x2, y1 - y2]
-
 const search = (initialState: State, data: DataStructure) => {
   data.add(initialState)
-
-  let boundaryNodesCount = 1
-  let visitedStatesCount = 0
-  let visitedNodes = 0
-  let maxBoundary = 0
-  let maxDepth = 0
 
   const visitedStates = new Set<Key>()
 
   do {
     const currentState = data.extract()
 
-    visitedStatesCount++
-    maxDepth = Math.max(currentState.depth, maxDepth)
-    maxBoundary = Math.max(data.size(), maxBoundary)
-
     if (isFinalState(currentState)) {
-      return {
-        state: currentState,
-        visitedStatesCount,
-        boundaryNodesCount,
-        maxDepth,
-        visitedNodes,
-        maxBoundary,
-      }
+      return currentState
     }
 
     if (visitedStates.has(currentState.key)) {
-      visitedNodes++
       continue
     }
     else {
@@ -163,35 +115,10 @@ const search = (initialState: State, data: DataStructure) => {
         }
 
         data.add({ ...nextState, cost: data.cost(nextState) })
-
-        boundaryNodesCount++
       }
     })
   } while (!data.isEmpty())
-  console.log(data.isEmpty(), { boundaryNodesCount })
   return null
-}
-
-// Final position in the grid for each number
-// 1 2 3  | [0,0] [0,1] [0,2]
-// 4 5 6  | [1,0] [1,1] [1,2]
-// 7 8 0  | [2,0] [2,1] [2,2]
-const finalPosition: Array<Pair> = [[2, 2]]
-for (let i = 0; i < 8; i++) {
-  finalPosition.push([Math.floor(i / 3), i % 3])
-}
-
-const manhattanDistance = (grid: Grid) => {
-  let sum = 0
-  for (let i = 0; i < MAX_BOUND; i++) {
-    for (let j = 0; j < MAX_BOUND; j++) {
-      const currentNumber = grid[i][j]
-      const finalPos = finalPosition[currentNumber]
-      const pairResult = pairDiff(finalPos, [i, j])
-      sum += (Math.abs(pairResult[0]) + Math.abs(pairResult[1]))
-    }
-  }
-  return sum
 }
 
 const compareStates = (sa: State, sb: State) => sa.cost < sb.cost
@@ -214,8 +141,11 @@ const makeAStarDS = (): DataStructure => ({
   cost: (state: State) => manhattanDistance(state.grid) + state.depth,
 })
 
+// Greedy
 export const Greedy = (initialState: State) =>
   perf(initialState, makeGreedyDS())
+
+// A Star
 export const AStar = (initialState: State) => perf(initialState, makeAStarDS())
 
 const perf = (initialState: State, dataStructure: DataStructure) => {
